feat: split liquid glass into plain and FFI demos
examples/liquid-glass is back to the pure GPUI setup (windowBackground
'blurred', no FFI); the NSGlassEffectView machinery moves to
examples/liquid-glass-ffi (demo:glass-ffi / bun:demo:glass-ffi). Both
render the same LiquidGlass.svelte — the FFI entry passes glass: true,
which also switches the caption. The FFI demo is deliberately not part
of the demo / bun:demo aggregates.
