+++ .gitignore
@@ -7,8 +7,8 @@ bun.lock
 # test:coverage working copy of svelte's sample suite
 test/.samples-tmp
 
-# compiled on demand by examples/liquid-glass/glass.js
-examples/liquid-glass/.glass-shim.dylib
+# compiled on demand by examples/liquid-glass-ffi/glass.js
+examples/liquid-glass-ffi/.glass-shim.dylib
 
 # output
 out
