+++ package.json
@@ -39,6 +39,7 @@
 		"demo:tictactoe": "node --conditions custom-renderer --conditions development --import ./src/register.js examples/tic-tac-toe/main.js",
 		"demo:hn": "node --conditions custom-renderer --conditions development --import ./src/register.js examples/hacker-news/main.js",
 		"demo:glass": "node --conditions custom-renderer --conditions development --import ./src/register.js examples/liquid-glass/main.js",
+		"demo:glass-ffi": "node --conditions custom-renderer --conditions development --import ./src/register.js examples/liquid-glass-ffi/main.js",
 		"test": "npm run test:reorder && npm run test:smoke",
 		"test:reorder": "node --conditions custom-renderer --conditions development --import ./src/register.js test/reorder.js",
 		"test:smoke": "node --conditions custom-renderer --conditions development --import ./src/register.js test/smoke.js",
@@ -48,6 +49,7 @@
 		"bun:demo:tictactoe": "bun --conditions custom-renderer --conditions development examples/tic-tac-toe/main.js",
 		"bun:demo:hn": "bun --conditions custom-renderer --conditions development examples/hacker-news/main.js",
 		"bun:demo:glass": "bun --conditions custom-renderer --conditions development examples/liquid-glass/main.js",
+		"bun:demo:glass-ffi": "bun --conditions custom-renderer --conditions development examples/liquid-glass-ffi/main.js",
 		"bun:test": "npm run bun:test:reorder && npm run bun:test:smoke",
 		"bun:test:reorder": "bun --conditions custom-renderer --conditions development test/reorder.js",
 		"bun:test:smoke": "bun --conditions custom-renderer --conditions development test/smoke.js",
